Updated phone number formatting for US/CA

+++ app/src/main/java/com/example/panikbutton/data/Contact.kt
@@ -3,6 +3,6 @@ package com.example.panikbutton.data
 data class Contact (
     val id: Long,
     val contactName: String,
-    val contactPhone: Int,
+    val contactPhone: Long,
     val contactEmail: String
 )
